pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gptoss-ollama"
version = "0.1.0"
description = "gptoss-ollama"
requires-python = ">=3.12,<3.13"
dependencies = [
  "openai>=1.99.9",
  "openai-harmony>=0.0.4",
  "openai-agents>=0.2.6",
  "langgraph>=0.6.5",
  "langgraph-prebuilt>=0.6.4",
  "langgraph-checkpoint>=2.1.1",
  "langchain>=0.3.27",
  "langchain-ollama>=0.3.6",
  "mcp>=1.12.4",
  "fastmcp>=2.11.3",
  "tavily-python>=0.7.10",
  "python-dotenv>=1.0.1",
  "pydantic>=2.11.7",
  "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
  "jupyter",
  "jupyterlab",
  "nbformat>=5.10,<6",
  "ipykernel",
]

[tool.setuptools.packages.find]
where = ["src"]